fix code inspection findings

+++ spring-swing-core/src/main/java/org/cosinus/swing/format/FormatHandler.java
@@ -154,7 +154,7 @@ public class FormatHandler {
             return "";
         }
         if (distance >= KILOMETER) {
-            return get2DecimalsFormatted((double) distance / KILOMETER) + " Km";
+            return get2DecimalsFormatted(distance / KILOMETER) + " Km";
         }
         return get2DecimalsFormatted(distance) + " m";
     }
